// src/components/ui/EiButton/EiButton.types.ts
import type { ButtonHTMLAttributes } from 'react';

export interface EiButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
    variant?: 'primary' | 'secondary' | 'outline' | 'ghost' | 'danger';
    size?: 'sm' | 'md' | 'lg';
    isLoading?: boolean;
    leftIcon?: React.ReactNode;
    rightIcon?: React.ReactNode;
    rounded?: boolean;
    fullWidth?: boolean;
}